+++ src/global.d.ts
@@ -1,0 +1,3 @@
+declare let stretchMode: boolean;
+declare let resizeCanvas: (informC?: boolean) => void;
+declare let getStartInstantly: () => boolean;