+++ src/fancade-plus.user.ts
@@ -1,5 +1,6 @@
 import { createManifestLink } from "./manifest";
 import { updateLocation, sanitizePath, setFullscreen, setInstantStart } from "./location";
+import { thanosSnap } from "./interface";
 
 document.head.append(createManifestLink(document));
 
@@ -8,3 +9,5 @@ sanitizePath(url);
 setInstantStart(url);
 setFullscreen(url);
 updateLocation(document.location, url);
+
+thanosSnap(document.getElementById("qc-cmp2-container")); // Yes, I would please like to take your cookies