+++ src/interface.ts
@@ -1,0 +1,3 @@
+export function thanosSnap(element: HTMLElement | null) {
+  element?.style.setProperty("visibility", "hidden", "important");
+}