+++ src/fancade-plus.user.ts
@@ -1,6 +1,7 @@
 import { createManifestLink } from "./manifest";
 import { updateLocation, sanitizePath, setFullscreen, setInstantStart } from "./location";
-import { thanosSnap } from "./interface";
+// @ts-expect-error bundler 
+import css from "./style.css"
 
 document.head.append(createManifestLink(document));
 
@@ -10,4 +11,6 @@ setInstantStart(url);
 setFullscreen(url);
 updateLocation(document.location, url);
 
-thanosSnap(document.getElementById("qc-cmp2-container")); // Yes, I would please like to take your cookies
+const style = document.createElement("style");
+style.innerHTML = css;
+document.head.append(style)