+++ src/interface.ts
@@ -1,3 +1,0 @@
-export function thanosSnap(element: HTMLElement | null) {
-  element?.style.setProperty("visibility", "hidden", "important");
-}