+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 ## [1.0.1](https://github.com/brckd/fancade-plus/compare/v1.0.0...v1.0.1) (2024-04-13)
 
-
 ### Bug Fixes
 
-* update dependenices ([fa0ee49](https://github.com/brckd/fancade-plus/commit/fa0ee491910c0705e032a168b528d23d3cb84dfa))
+- update dependenices
+  ([fa0ee49](https://github.com/brckd/fancade-plus/commit/fa0ee491910c0705e032a168b528d23d3cb84dfa))
 
 # 1.0.0 (2024-04-13)