+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # [1.1.0](https://github.com/brckd/fancade-plus/compare/v1.0.1...v1.1.0) (2024-04-13)
 
-
 ### Features
 
-* modify params for all windows ([094d897](https://github.com/brckd/fancade-plus/commit/094d8970495ccbdbe267492f57ef81195b202ad8))
+- modify params for all windows
+  ([094d897](https://github.com/brckd/fancade-plus/commit/094d8970495ccbdbe267492f57ef81195b202ad8))
 
 ## [1.0.1](https://github.com/brckd/fancade-plus/compare/v1.0.0...v1.0.1) (2024-04-13)